requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-ollama==0.0.1
faiss-cpu==1.7.4
pypdf==3.17.4
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.18
ollama==0.1.7
